Add files via upload
Added my first Python file to the repository, ProtoDiary1 - the first step in my Personal Diary Project. This file takes user input in a few fields (title, date, time, then the main text body) and creates a .txt file in a text diary entry format. The .txt file is created in whichever folder the program is running in, but in the future I hope to be able to direct the .txt to be saved in a folder of the user's choosing.