Started ProtoDiary2
Started work on ProtoDiary2, which will be a text diary entry program similar to PD1, but instead of producing a .txt file it will save the entry to a JSON file that can be viewed in the Python program itself. Have added a "new entry" function, which is very unpolished at the moment but it at least seems to successfully save the data. Next steps are to allow multiple entries to be saved at a time, and also to make the console look a little less messy when using the program.