Added "New Entry" page
Started work on the second page, the "New Entry" page. Tidied the code up a bit and commented it somewhat, and was also able to activate the "New Entry" button to load the new page when pressed. The program is starting to come together nicely and I'm gaining an understanding of how things are done with Python Tkinter.

Next steps are to fix issues with the new page - the textboxes (for entering new text) are a bit broken, and I also need to add a submit button and activate it to allow it to take that text and add it to a JSON file. Fun times ahead!