"New Entry" save to .txt file functionality added
Have added some functionality to the "New Entry" page. Have started off by simply saving an entry to a .txt file, which after much trial and error I managed to get it to do. Next step is getting JSON set up and allowing entries to be saved to that, as well as starting work on the other pages to allow use to access the saved entries.