Started work on "View Saved Entries" page
The "View Saved Entries" page is now implemented onto the program. The page is built and designed, and the "View Saved Entries" button on the main menu successfully navigates to it, and the "Home" button on the new page successfully navigates back to home. The design looks as it should from what I wanted.

Next steps, JSON is imported so will refactor the "submit" button to save the new entry there instead of as a .txt file, then it will be time to add a function that adds the entry to the "view saved" page as a clickable button that loads a page from which to read the entry.