Added "Read Entry" and "Delete Entry" pages, added read entry functionality
I have managed to add a "Read Entry" page as well as a way for the user to select an article from the "View Entries" page to read (albeit not in the way I wanted - have implemented a system of typing in a name and hitting a "submit" button that takes the user to the "read" page for that article, as opposed to a button system as I wanted).

Also started work on a "Delete Entry" page. The page is built but not functional at the moment. Building functionality for that and building the "Help" page will be the next goals.