Added "Delete Entry" page functionality
Added functionality the the "Delete Entry" page. It proved a more difficult task than anticipated, as I had to refactor the code for the page into a function as well as nest the other delete page functions into it. I had to do this as when using the delete page, it didn't refresh properly - if the user visited the page and deleted an entry, if they then left and came back, the entry would still be displayed along with the "deleted successfully" message.

This issue is now resolved, but I'm definitely going to need to spend some time going over the code. Everything is working roughly as it should, but after shifting a lot of things around I'll need to review my code to get a full understanding of where everything is.