Set up the "Help & Readme" page
Have created the "Help & Readme" page with some simple instructions on how to use the program. This marks the completion of all of the program's basic features.

However, there is still much work to be done. The current state of the program code and certain aspects of the program itself are, to put it bluntly, an unholy mess that manages to avoid tripping over itself enough to somewhat work. I'm going to fix and clean up this version of the program, and will also likely copy this over to another prototype (PDGUI-V2) to start working on different ways to make the program look and function a little bit better.