Created all pages of program, added functionality to "New Entry" and "View Saved Entries" pages
Have been able to leverage my experience with PDGUI-V1 to rapidly build GUI-V2. All pages are built and the navigation buttons are linked, and the "New Entry" and "View Saved Entries" (and my extension the "Read Saved Entry") pages are all functional. 

The "View Saved Entries" page now functions with the entries as buttons that lead directly to the "Read Entry" page - this is what I always wanted to do, but it took me a good deal of thought as to how to do it (it was a lack of knowledge of how to implement the "lambda" functions that tripped me up - but I've learned something through it).

Overall, the code for GUI-V2 is much cleaner than GUI-V1. Very happy with it so far. Next step, get the "Delete" and "Help" pages up and running!