Finishing touches to PDGUI-V2
Added some finishing touches to the second GUI prototype, adding a "word wrap" to the big text entry and reading fields to prevent words from splitting during line breaks. Folded all of the home page to other page functions into one. Made the "delete entry" page labels a bit tidier, now displaying the date as well as the title but making one clear from the other.

This marks the end of the second version of the prototype. There are a few features that I would like to try to implement but will likely copy what I have over to a "PDGUI-V2.5" file and experiment from there whilst leaving this working one intact.