PDGUI-V3 Underway
Having made GUI-V2 about as good as I can think to make it as a text diary, I've decided to copy what I've done there into a new file and start PDGUI-V3 to experiment with. I've already made a system for the "Delete Entry" page that utilize a button rather than a textbox and submit button for deletions (although I do wish to refine this further so that accidental deletions are less likely). 

I've also decided to switch up the theme for this one, making it a black/orange/yellow "honeycomb" theme to make it distinct from V2, although I'm still in the process of switching the colours over on all of the pages. 

Finally, I've refactored the "Home Page" to include a button for an "Audio Entry" page. This will be the big experiment of Version 3 - branching out into different media types in my pursuit to make a full-spectrum multimedia diary app with Python!