Completed theme change, started "Audio Entries" page
Have completed the transition from the "aero" theme (blues and greens) to the "honeycomb" theme (black and orange). It's given me another idea - I want to add an option to the program that allows the user to switch themes to one that they like. Will work on this soon.

Also started work on the "Audio Entries" page. Links working and the basic setup is there but my next challenge will be implementing the functionality of it.