Added functions for displaying and "loading" audio entries
Added some functionality to the "Audio Entries" page, including a "display audio entries" function which loops through a folder titled "Audio Entries" where the .mp3 files for the audio entries are kept. This function then displays each of these as buttons, similar to the "View Entries" page.

This then leads to the "Load Audio Entry" function, which loads the chosen entry into the audio interface when the button for that entry is pressed.

Next step: make the audio interface work, able to play and stop the chosen entry using the buttons available.