Imported Pygame and activated audio interface
Managed to import the "pygame" module along with the "mixer" element and used this to activate my audio interface to play my stored audio entries. The user can now click a button for their chosen entry to "load" it into the player, then press "play" on the player to play it, and "stop" to stop playback.

The logical next step would be to either have an in-built way of recording new audio entries, or a way of having a button that opened the "Audio Entries" folder allowing a user to add their pre-recorded .mp3 files to it it. Other features, such as a volume slider and "pause" button should also be looked at.