Added my "Audio Entries" folder two entries
Added an "Audio Entries" folder with a couple of (somewhat terrible) pre-recorded entries for diagnostic purposes. This should be kept in the same folder as the program in order to be compatible with the file path in the program - otherwise, the program won't be able to find and use the files.