Added more "Exit" buttons, started work on additional audio features
Added an "exit" button to each of the other pages, so the user can now exit the program from any page, and also started work on the volume system for the audio interface, creating the buttons and label. Very unsure if what I have in my head is going to work, but will give it a good go nonetheless.