Started work on "Add Audio Entry" page, added pause and unpause buttons to audio interface
Added pause and unpause buttons to the audio interface, and added functionality to both. I initially tried using the single button for both, as it is done pretty much everywhere else, but it's proved a tricky problem so I used two buttons purely to make sure that the pause functionality would work. A problem to solve in the future for sure.

I also started work on an "Add Audio Entry" page. The design has been prototyped, so to speak, although I'll likely add to it. Functionality is still to be added as well.