Added functionality to "add audio entry" button
Added functionality to the part of the "Add Audio Entries" page that allows a user to add an existing .mp3 file to the "Audio Entries" folder. I was able to record an entry, then go into the app, press the button, have a file dialog window pop up set at the folder, and copy/drag and drop the new entry in.

Also started work on the "record entry" part, building an entry for entering a duration in seconds (since the recording function has to be pre-set with a duration to record for).