Integrated audio recording now working
The "record audio entry" part of the add audio entries page is now operational. Actually getting the audio recording to work in the first instance was actually relatively easy, but it was converting from the .wav format to a .mp3 file that the program actually uses that proved tricky.

The trick was that I needed to download ffmpeg and add it to my system's PATH. Even then, my program really didn't seem to want to recognise it, but after a significant amount of experimentation including specifying the path in the program itself as well as specifying a path for the .mp3 file to go, it seemed to work eventually. What ultimately went right will need to be looked into.

Next steps will be to add a countdown timer to the recorder so the user can see how much time they have left to talk, as well as add dates to the view audio entries so they don't end up out of order.