Started work on PDGUI-V4
Having successfully implemented audio entries into the program, I thought it was time to get to work on Version 4. My goals for this version are to get image and video entries up and running as well as to make improvements to the audio entries pages. The theme of this version will be "Mystery" to reflect my enjoyment of a recent TV show.