Continued with thematic changes for GUI-V4
Implemented the theme change that I had made for the home page onto the rest of the pages. Will likely change a few other things but overall I am really happy with it!

Next steps will be to set up the image entries and video entries pages, although having done a review of the full program whilst changing the themes over, I have a few ideas for refinements that I'll likely make along the way.